CHORE Moved InstantiateMacro to MacroNode
