+++ Assets/com.alelievr.NodeGraphProcessor/Runtime/Elements/MacroNode.cs
@@ -1,5 +1,6 @@
 using System;
 using GraphProcessor;
+using UnityEngine;
 
 [System.Serializable]
 public class MacroNode : SubGraphNode
@@ -9,5 +10,13 @@ public class MacroNode : SubGraphNode
     public void SetMacro(SubGraph macro)
     {
         this.subGraph = macro;
+    }
+
+    public static BaseNode InstantiateMacro(Type nodeType, Vector2 position, params object[] args)
+    {
+        SubGraph macro = args[0] as SubGraph;
+        MacroNode macroNode = BaseNode.CreateFromType(nodeType, position, args) as MacroNode;
+        macroNode.SetMacro(macro);
+        return macroNode;
     }
 }
